+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ### 1.2.0 
+- AND: Updated to jetpack/androidx
 - AND: Updated to FreKotlin 1.9.1
 - iOS: Updated to FreSwift 4.1.0
-- iOS: upgrade to 5.11.1 FBSDK
+- iOS: upgrade to 5.13.1 FBSDK
 - iOS: use Swift Package Manager
 - iOS: MessageDialog is now deprecated
-- Android: upgrade to 5.11.2 FBSDK
+- Android: upgrade to 5.13.0 FBSDK
 
 ### 1.1.0 
 - add `FacebookSdk.onShareSuccess`